# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/login": {
            "post": {
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "5d4cdb9e3d5c21bb811c6fc86d43d291",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/admin/profile": {
            "get": {
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Get profile user details",
                "operationId": "539f1e7bcbb56775eb8ade21b35d2a96",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/logout": {
            "post": {
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Logout",
                "operationId": "e57bd33e5159eb824d6c2f4e75b8dfa8",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/change-profile": {
            "post": {
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Change Profile information",
                "operationId": "fa4bb2978d26def13fdf3a52a3025ead",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "old_password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "example": "123123"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Hoàng Anh"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0342835419"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/categories/index": {
            "get": {
                "tags": [
                    "Categories Management"
                ],
                "summary": "Get list of categories",
                "operationId": "81d5c63cfe0e75c2c4577a0600b8308a",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search categories by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "food"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for paginated results",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/categories/detail": {
            "get": {
                "tags": [
                    "Categories Management"
                ],
                "summary": "Get detail of category",
                "operationId": "34aea717e23e806a24b10c3f3df32172",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/categories/create": {
            "post": {
                "tags": [
                    "Categories Management"
                ],
                "summary": "Create a new category",
                "operationId": "9e390035ddbb83b40164f08d1ea49f22",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Hoàng Anh"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/categories/edit": {
            "post": {
                "tags": [
                    "Categories Management"
                ],
                "summary": "Update new category",
                "operationId": "a8d7d75e63779b99cfc031f3368178d2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Hoàng Anh"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/categories/delete": {
            "delete": {
                "tags": [
                    "Categories Management"
                ],
                "summary": "Delete a category",
                "operationId": "6f752e19c758ec2a0cb684824f644a40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Category ID is deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}